Split JetStream read and write adapters
