+++ tests/NatsManager.Web.Tests/NatsManagerWebAppFactory.cs
@@ -31,7 +31,7 @@ public sealed class NatsManagerWebAppFactory : WebApplicationFactory<Program>
     private readonly SqliteConnection _connection = new("DataSource=:memory:");
 
     public IEnvironmentRepository EnvironmentRepository { get; } = Substitute.For<IEnvironmentRepository>();
-    public IJetStreamAdapter JetStreamAdapter { get; } = Substitute.For<IJetStreamAdapter>();
+    public IJetStreamAdapter JetStreamReadAdapter { get; } = Substitute.For<IJetStreamAdapter>();
     public IJetStreamWriteAdapter JetStreamWriteAdapter { get; } = Substitute.For<IJetStreamWriteAdapter>();
     public IKvStoreAdapter KvStoreAdapter { get; } = Substitute.For<IKvStoreAdapter>();
     public IObjectStoreAdapter ObjectStoreAdapter { get; } = Substitute.For<IObjectStoreAdapter>();
@@ -88,7 +88,7 @@ public sealed class NatsManagerWebAppFactory : WebApplicationFactory<Program>
 
             // Replace NATS adapters with substitutes
             ReplaceService<IEnvironmentRepository>(services, EnvironmentRepository);
-            ReplaceService<IJetStreamAdapter>(services, JetStreamAdapter);
+            ReplaceService<IJetStreamAdapter>(services, JetStreamReadAdapter);
             ReplaceService<IJetStreamWriteAdapter>(services, JetStreamWriteAdapter);
             ReplaceService<IKvStoreAdapter>(services, KvStoreAdapter);
             ReplaceService<IObjectStoreAdapter>(services, ObjectStoreAdapter);
